requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
cssselect==1.2.0
fake-useragent==1.3.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
multidict==6.0.4
nextcord==2.6.0
nextcord-pretty-help==1.3.3
parse==1.19.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dotenv==1.0.0
requests==2.31.0
requests-html==0.10.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.8.0
Unidecode==1.3.7
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
yarl==1.9.2
zipp==3.17.0
